Refactore code. Moved add_classification_job to ustiles file. Also moved function allowed_file to new utils file.
